requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
certifi==2025.6.15
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
prisma==0.15.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.2.0
sniffio==1.3.1
starlette==0.46.2
StrEnum==0.4.15
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
